# Optimize DecodeMapToType