--- a/MaxMind.Db/Decoder.cs
+++ b/MaxMind.Db/Decoder.cs
@@ -1,7 +1,7 @@
 ﻿#region
 
 using System;
-#if NETSTANDARD2_1_OR_GREATER || NET8_0_OR_GREATER
+#if !NETSTANDARD2_0
 using System.Buffers;
 #endif
 using System.Collections;
@@ -351,7 +351,7 @@ namespace MaxMind.Db
         {
             var constructor = _typeAcivatorCreator.GetActivator(expectedType);
 
-#if NETSTANDARD2_1_OR_GREATER || NET8_0_OR_GREATER
+#if !NETSTANDARD2_0
             // N.B. Rent can return a larger arrays. This is fine because constructors allow arrays larger than the
             // number of parameters.
             object?[] parameters = ArrayPool<object?>.Shared.Rent(constructor.DefaultParameters.Length);
@@ -383,7 +383,7 @@ namespace MaxMind.Db
             outOffset = offset;
             object obj = constructor.Activator(parameters);
 
-#if NETSTANDARD2_1_OR_GREATER || NET8_0_OR_GREATER
+#if !NETSTANDARD2_0
             ArrayPool<object?>.Shared.Return(parameters);
 #endif
 
@@ -403,7 +403,7 @@ namespace MaxMind.Db
 
                 var activator = _typeAcivatorCreator.GetActivator(param.ParameterType);
 
-#if NETSTANDARD2_1_OR_GREATER || NET8_0_OR_GREATER
+#if !NETSTANDARD2_0
                 object?[] cstorParams = ArrayPool<object?>.Shared.Rent(activator.DefaultParameters.Length);
 #else
                 object?[] cstorParams = new object?[activator.DefaultParameters.Length];
@@ -415,7 +415,7 @@ namespace MaxMind.Db
                 SetAlwaysCreatedParams(activator, cstorParams, injectables, network);
                 parameters[param.Position] = activator.Activator(cstorParams);
 
-#if NETSTANDARD2_1_OR_GREATER || NET8_0_OR_GREATER
+#if !NETSTANDARD2_0
                 ArrayPool<object?>.Shared.Return(cstorParams);
 #endif
             }